strands-hands-on/pyproject.toml
[project]
name = "strands-hands-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp-server-git>=2025.11.25",
    "mcp[cli]>=1.22.0",
    "strands-agents-tools>=0.2.16",
    "strands-agents[gemini]>=1.18.0",
]
